refac: Sidebar 리팩토링
